+++ app/controllers/exercises_controller.rb
@@ -356,7 +356,7 @@ class ExercisesController < ApplicationController
 
     return handle_proforma_multi_import(result) if result.is_a?(Array)
 
-    redirect_to edit_exercise_path(result), notice: t('controllers.exercise.import_proforma_xml.single_import_successful')
+    redirect_to result, notice: t('controllers.exercise.import_proforma_xml.single_import_successful')
   rescue Proforma::PreImportValidationError
     redirect_to exercises_path, alert: t('controllers.exercise.import_proforma_xml.validation_error')
   end
