+++ app/services/proforma_service/convert_task_to_exercise.rb
@@ -1,0 +1,87 @@
+# frozen_string_literal: true
+
+module ProformaService
+  class ConvertTaskToExercise < ServiceBase
+    def initialize(task: nil, user: nil, exercise: nil)
+      @task = task
+      @user = user
+      @exercise = exercise || Exercise.new
+    end
+
+    def execute
+      import_exercise
+      @exercise
+    end
+
+    private
+
+    def import_exercise
+      @exercise.assign_attributes(
+        uuid: @task.uuid,
+        user: @user,
+        title: @task.title,
+        private: true,
+        descriptions: descriptions,
+        instruction: @task.internal_description,
+        execution_environment: execution_environment,
+        tests: tests,
+        exercise_files: task_files.values,
+        tag_list: @exercise.persisted? ? 'updated' : 'new'
+      )
+    end
+
+    def descriptions
+      @exercise.descriptions[0] = [Description.new(text: @task.description, language: @task.language)]
+    end
+
+    def task_files
+      @task_files ||= Hash[
+        @task.all_files.reject { |file| file.id == 'ms-placeholder-file' }.map do |task_file|
+          [task_file.id, exercise_file_from_task_file(task_file)]
+        end
+      ]
+    end
+
+    def exercise_file_from_task_file(task_file)
+      ExerciseFile.new({
+        full_file_name: task_file.filename,
+        read_only: task_file.usage_by_lms.in?(%w[display download]),
+        hidden: task_file.visible == 'no',
+        role: task_file.internal_description
+      }.tap do |params|
+        if task_file.binary
+          params[:attachment] = file_base64(task_file)
+          params[:attachment_file_name] = task_file.filename
+          params[:attachment_content_type] = task_file.mimetype
+        else
+          params[:content] = task_file.content
+        end
+      end)
+    end
+
+    def file_base64(file)
+      "data:#{file.mimetype || 'image/jpeg'};base64,#{Base64.encode64(file.content)}"
+    end
+
+    def tests
+      @task.tests.map do |test_object|
+        Test.new(
+          feedback_message: test_object.meta_data['feedback-message'],
+          testing_framework: TestingFramework.where(
+            name: test_object.meta_data['testing-framework'],
+            version: test_object.meta_data['testing-framework-version']
+          ).first_or_initialize,
+          exercise_file: test_file(test_object)
+        )
+      end
+    end
+
+    def test_file(test_object)
+      task_files.delete(test_object.files.first.id).tap { |file| file.purpose = 'test' }
+    end
+
+    def execution_environment
+      ExecutionEnvironment.where(language: @task.proglang[:name], version: @task.proglang[:version]).first_or_initialize
+    end
+  end
+end
