+++ app/services/proforma_service/import.rb
@@ -10,8 +10,10 @@ module ProformaService
     def execute
       if single_task?
         importer = Proforma::Importer.new(@zip)
-        @task = importer.perform
-        import_exercise
+        task = importer.perform
+        exercise = ConvertTaskToExercise.call(task: task, user: @user, exercise: Exercise.find_by(uuid: task.uuid))
+        exercise.save!
+        exercise
       else
         import_multi
       end
@@ -26,10 +28,9 @@ module ProformaService
           zip_files.map! do |entry|
             store_zip_entry_in_tempfile entry
           end
-          exercises = zip_files.map do |proforma_file|
+          zip_files.map do |proforma_file|
             Import.call(zip: proforma_file, user: @user)
           end
-          exercises.each(&:save)
         ensure
           zip_files.each(&:unlink)
         end
@@ -49,71 +50,6 @@ module ProformaService
       end
 
       filenames.select { |f| f[/\.xml$/] }.any?
-    end
-
-    def import_exercise
-      @exercise = Exercise.create(
-        uuid: @task.uuid,
-        user: @user,
-        title: @task.title,
-        private: true,
-        descriptions: [Description.new(text: @task.description, language: @task.language)],
-        instruction: @task.internal_description,
-        execution_environment: execution_environment,
-        tests: tests,
-        exercise_files: task_files.values,
-        tag_list: 'new'
-      )
-    end
-
-    def task_files
-      @task_files ||= Hash[
-        @task.all_files.reject { |file| file.id == 'ms-placeholder-file' }.map do |task_file|
-          [task_file.id, exercise_file_from_task_file(task_file)]
-        end
-      ]
-    end
-
-    def exercise_file_from_task_file(task_file)
-      ExerciseFile.new({
-        full_file_name: task_file.filename,
-        read_only: task_file.usage_by_lms.in?(%w[display download]),
-        hidden: task_file.visible == 'no',
-        role: task_file.internal_description
-      }.tap do |params|
-        if task_file.binary
-          params[:attachment] = file_base64(task_file)
-          params[:attachment_file_name] = task_file.filename
-          params[:attachment_content_type] = task_file.mimetype
-        else
-          params[:content] = task_file.content
-        end
-      end)
-    end
-
-    def file_base64(file)
-      "data:#{file.mimetype || 'image/jpeg'};base64,#{Base64.encode64(file.content)}"
-    end
-
-    def tests
-      @task.tests.map do |test_object|
-        Test.new(
-          feedback_message: test_object.meta_data['feedback-message'],
-          testing_framework: TestingFramework.where(
-            name: test_object.meta_data['testing-framework'],
-            version: test_object.meta_data['testing-framework-version']
-          ).first_or_initialize,
-          exercise_file: test_file(test_object)
-        )
-      end
-    end
-
-    def test_file(test_object)
-      task_files.delete(test_object.files.first.id).tap { |file| file.purpose = 'test' }
-    end
-
-    def execution_environment
-      ExecutionEnvironment.where(language: @task.proglang[:name], version: @task.proglang[:version]).first_or_initialize
     end
   end
 end
