+++ db/migrate/20190731170546_add_uniqueness_index_to_exercise_uuid.rb
@@ -1,0 +1,5 @@
+class AddUniquenessIndexToExerciseUuid < ActiveRecord::Migration[5.2]
+  def change
+    add_index :exercises, :uuid, unique: true
+  end
+end
