+++ db/migrate/20190801135856_add_primary_to_descriptions.rb
@@ -1,0 +1,5 @@
+class AddPrimaryToDescriptions < ActiveRecord::Migration[5.2]
+  def change
+    add_column :descriptions, :primary, :boolean
+  end
+end
